+++ src/components/level/sentry.ts
@@ -30,6 +30,18 @@ export interface SentryComp extends Comp {
      * Binds the event fired when objects of interest are spotted.
      */
     onObjectsSpotted(cb: (objects: GameObj[]) => void): EventController;
+    /*
+     * Returns whether the object is within the field of view.
+     */
+    isWithinFieldOfView(
+        obj: GameObj<PosComp>,
+        direction?: Vec2,
+        fieldOfView?: number,
+    ): boolean;
+    /*
+     * Returns whether there is a line of sight to the object.
+     */
+    hasLineOfSight(obj: GameObj<PosComp>): boolean;
 }
 
 /**
@@ -74,6 +86,9 @@ export function sentry(
             return k.query(candidates);
         };
     const checkFrequency = opts.checkFrequency || 1;
+    const directionVector = typeof opts.direction === "number"
+        ? Vec2.fromAngle(opts.direction)
+        : opts.direction;
     let t = 0;
     return {
         id: "sentry",
@@ -90,7 +105,33 @@ export function sentry(
         get directionAngle(): number {
             return this.direction ? this.direction.angle() : undefined;
         },
-        fieldOfView: opts.fieldOfView || 200,
+        fieldOfView: opts.fieldOfView || 200, // 200 degrees = Human field of view
+        isWithinFieldOfView(
+            this: GameObj<SentryComp | PosComp>,
+            obj: GameObj<PosComp>,
+            direction?: Vec2,
+            fieldOfView?: number,
+        ) {
+            const dir: Vec2 = (typeof direction === "number"
+                ? Vec2.fromAngle(direction)
+                : direction) || directionVector;
+            const fov: number = fieldOfView || opts.fieldOfView;
+            if (!dir || !fov || fov >= 360) return true;
+            const halfAngle = fov / 2;
+            return obj.pos
+                && dir.angleBetween(obj.pos.sub(this.pos)) <= halfAngle;
+        },
+        hasLineOfSight(
+            this: GameObj<SentryComp | PosComp>,
+            obj: GameObj<PosComp>,
+        ) {
+            const hit = raycast(
+                this.pos,
+                obj.pos.sub(this.pos),
+                opts.raycastExclude,
+            );
+            return hit && hit.object === obj;
+        },
         update(this: GameObj<SentryComp | PosComp>) {
             t += k.dt();
             if (t > checkFrequency) {
@@ -98,25 +139,21 @@ export function sentry(
                 let objects = get();
                 // If fieldOfView is used, keep only object within view
                 if (
-                    objects.length && this.direction && this.fieldOfView
+                    objects.length && directionVector && this.fieldOfView
                     && this.fieldOfView < 360
                 ) {
                     const halfAngle = this.fieldOfView / 2;
                     objects = objects.filter(o =>
                         o.pos
-                        && this.direction.angleBetween(o.pos.sub(this.pos))
+                        && directionVector.angleBetween(o.pos.sub(this.pos))
                             <= halfAngle
                     );
                 }
                 // If lineOfSight is used, raycast
                 if (objects.length && opts.lineOfSight) {
                     objects = objects.filter(o => {
-                        const hit = raycast(
-                            this.pos,
-                            o.pos.sub(this.pos),
-                            opts.raycastExclude,
-                        );
-                        return hit.object === o;
+                        return o.pos
+                            && this.hasLineOfSight(o as GameObj<PosComp>);
                     });
                 }
                 if (objects.length > 0) {